+++ packages/cli-adapter/src/saihubot-cli-skill-help.js
@@ -16,13 +16,11 @@ export const skillHelp = {
       return skill.help ? `* ${skill.name}\n  ${skill.help}` :
         `* ${skill.name}`;
     });
-    // const pre = document.createElement('pre');
-    const content = (
-      <Box borderStyle="round" borderColor="cyan" float="center" padding={1}>
-        {`I have ${availableSkills.length} skills:\n ${availableSkills.join('\n')}`}
-      </Box>
+    robot.adapter.unsafe_sendComponent(
+        <Box borderStyle="round" borderColor="cyan" float="center" padding={1}>
+          {`I have ${availableSkills.length} skills:\n ${availableSkills.join('\n')}`}
+        </Box>
     );
-    robot.adapter.send(content);
   },
 };